+++ src/main/kotlin/io/kjson/JSONLong.kt
@@ -50,6 +50,18 @@ class JSONLong(override val value: Long) : JSONNumberValue(), JSONValue {
 
     override fun isByte(): Boolean = value in Byte.MIN_VALUE..Byte.MAX_VALUE
 
+    override fun isZero(): Boolean = value == 0L
+
+    override fun isNegative(): Boolean = value < 0
+
+    override fun isPositive(): Boolean = value > 0
+
+    override fun isNotZero(): Boolean = value != 0L
+
+    override fun isNotNegative(): Boolean = value >= 0
+
+    override fun isNotPositive(): Boolean = value <= 0
+
     override fun toDouble(): Double = value.toDouble()
 
     override fun toFloat(): Float = value.toFloat()